Improve understandability of spoken instructions.
The speech system doesn't do a superb job of reading system names. This
commit includes a couple changes to make it more understandable:
- lower speechSynthesizer.Rate to -2;
- split large numbers into individual digits;
- give the light year distance of each jump.

The last is most helpful, as the list of systems is sorted by light
year.

Also: use MessageBox.Show to report errors when setting the log
directory.